+++ src/settings/IdentifierTypesSettings.js
@@ -22,6 +22,8 @@ class IdentifierTypesSettings extends React.Component {
   suppressDelete = term => term.source === 'folio';
 
   render() {
+    const hasPerm = this.props.stripes.hasPerm('ui-inventory.settings.identifier-types');
+
     return (
       <IntlConsumer>
         {intl => (
@@ -45,6 +47,7 @@ class IdentifierTypesSettings extends React.Component {
             actionSuppressor={{ edit: this.suppressEdit, delete: this.suppressDelete }}
             id="identifier-types"
             sortby="name"
+            editable={hasPerm}
           />
         )}
       </IntlConsumer>