+++ src/settings/InstanceStatusTypesSettings.js
@@ -24,6 +24,8 @@ class InstanceStatusTypesSettings extends React.Component {
   suppressDelete = term => term.source === 'marcrelator';
 
   render() {
+    const hasPerm = this.props.stripes.hasPerm('ui-inventory.settings.instance-statuses');
+
     return (
       <IntlConsumer>
         {intl => (
@@ -48,6 +50,7 @@ class InstanceStatusTypesSettings extends React.Component {
             id="instanceStatus-types"
             sortby="name"
             validate={validateNameAndCode}
+            editable={hasPerm}
           />
         )}
       </IntlConsumer>