+++ test/bigtest/interactors/instance-view-page.js
@@ -45,7 +45,6 @@ import MultiColumnListInteractor from '@folio/stripes-components/lib/MultiColumn
   hasHeaderDropdown = isPresent('[class^=DropdownMenuTether---]');
   headerDropdownMenu = new HeaderDropdownMenu();
   itemsList = new MultiColumnListInteractor('#list-items');
-  instanceNotesList = new MultiColumnListInteractor('#list-instance-notes');
   seriesStatementList = new MultiColumnListInteractor('#list-series-statement');
   subjectsList = new MultiColumnListInteractor('#list-subject');
   items = collection('#list-items div[class^=mclRow]', Item);