+++ test/bigtest/interactors/inventory.js
@@ -12,6 +12,16 @@ export default @interactor class InventoryInteractor {
   static defaultScope = '[data-test-inventory-instances]';
 
   isNoResultsMessageLabelPresent = isPresent('[class*=noResultsMessageLabel]');
+
+  isResourceFilterPresent = isPresent('[data-test-filter-instance-resource]');
+  isLocationFilterPresent = isPresent('[data-test-filter-instance-location]');
+  isStaffSuppressFilterPresent = isPresent('[data-test-filter-instance-staff-suppress]');
+  isDiscoverySuppressFilterPresent = isPresent('[data-test-filter-instance-discovery-suppress]');
+  clickSelectStaffSuppressFilter = clickable('#clickable-filter-staffSuppress-true');
+  clickClearStaffSuppressFilter = clickable('#staffSuppress button[class^="iconButton---"]');
+  clickSelectDiscoverySuppressFilter = clickable('#clickable-filter-discoverySuppress-true');
+  clickClearDiscoverySuppressFilter = clickable('#discoverySuppress button[class^="iconButton---"]');
+
   instances = collection('#list-inventory [role=row] a');
 
   instance = scoped('[data-test-instance-details]');