+++ test/bigtest/interactors/settings/resource-types/instance-types.js
@@ -7,6 +7,7 @@ import {
 @interactor class InstanceTypes {
   hasList = isPresent('#editList-instance-types');
   rowCount = count('[class^="editListRow---"]');
+  hasNewButton = isPresent('#clickable-add-instance-types');
 }
 
 export default new InstanceTypes();